- added basic QuadTree implementation (WIP) - added WorldMath.natvis file, a sample of what is yet to come! - added boolean ConsoleArgumentType implementation - fixed LoopRange * dereference operator (now returns actual type instead of int ptr) - added math self-test to before running ATG - added add_cdimage command because it was missing - disabled some stupid warnings regarding deprecated C++/C features - made some progress on WorldMath.cpp plane equations (not done yet) - fixed some compiler warnings by casting things properly

+++ krt/streaming/src/CdImageDevice.cpp
@@ -90,7 +90,8 @@ bool CdImageDevice::OpenImage(const std::string& imagePath)
 		}
 
 		// use the directory path in case bulk handles are equivalent to the parent handle
-		numEntries = parentDevice->GetLength(directoryPath) / sizeof(Entry);
+        // We assume that you dont actually be that crazy and make such tiny entires in the CdImage that you have more than 2^32 - 1.
+		numEntries = (uint32_t)( parentDevice->GetLength(directoryPath) / sizeof(Entry) );
 	}
 
 	// read the directory entry list
